refactored module
